Add files via upload
Jarvis Mark 1 is the basic AI desktop voice assistant which takes the audio commands as input and performs the simple tasks based on the commands given. It can play music, play videos on youtube and it also can do google, wikipedia searches. In this version( Mark 1 ) less functionalities are defined. In future work of the other versions we will add more functionalities to Jarvis.